+++ src/constants/cardTypes.ts
@@ -1,0 +1,10 @@
+export const CARD_TYPES: Record<string, { name: string; color: string }> = {
+  RED: { name: '찬욱 카드', color: '#E24141' },
+  BLUE: { name: '효리 카드', color: '#547CE4' },
+  GREEN: { name: '수연 카드', color: '#73BC6D' },
+  PINK: { name: '세진 카드', color: '#DE59B9' },
+  MINT: { name: '진경 카드', color: '#94DACD' },
+  CORAL: { name: '종길 카드', color: '#E76E9A' },
+  ORANGE: { name: '건우 카드', color: '#F37D3B' },
+  YELLOW: { name: '혜성 카드', color: '#FBCD58' },
+}
